Working on Vinyl class

+++ VStepanov.Experiments.Vinyl.Imaging/Image.cs
@@ -5,11 +5,16 @@ namespace VStepanov.Experiments.Vinyl.Imaging
 {
     public class Image
     {
-        private byte[,] _imageData;
+        #region Fields
+        private int[,] _imageData; 
+        #endregion
 
+        #region Properties
         public int Width { get; private set; }
-        public int Height { get; private set; }
+        public int Height { get; private set; } 
+        #endregion
 
+        #region Constructors and constructing methods
         /// <summary>
         /// Class, providing access to image data.
         /// 
@@ -18,10 +23,10 @@ namespace VStepanov.Experiments.Vinyl.Imaging
         /// <param name="bitmap">Source image.</param>
         public Image(Bitmap bitmap)
         {
-            Width  = bitmap.Width;
+            Width = bitmap.Width;
             Height = bitmap.Height;
 
-            _imageData = new byte[Width, Height];
+            _imageData = new int[Width, Height];
 
             for (int x = 0; x < Width; x++)
             {
@@ -32,12 +37,18 @@ namespace VStepanov.Experiments.Vinyl.Imaging
             }
         }
 
+        /// <summary>
+        /// Creates an instance of Image class from selected file.
+        /// </summary>
+        /// <param name="path">Path to file with image.</param>
+        /// <returns></returns>
         public static Image FromFile(string path)
         {
             var bitmap = new Bitmap(path);
 
             return new Image(bitmap);
-        }
+        } 
+        #endregion
 
         /// <summary>
         /// Returns lightness of desired pixel.
@@ -45,7 +56,7 @@ namespace VStepanov.Experiments.Vinyl.Imaging
         /// <param name="x">X-coordinate</param>
         /// <param name="y">Y-coordinate</param>
         /// <returns>Lightness at [x, y]</returns>
-        public byte this[int x, int y]
+        public int this[int x, int y]
         {
             get
             {
